# src/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Projeto Lanchonete | Tech Challenge",
    "description": "API Documentation",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
...
    "/order": {
      "get": {
        "tags": ["Orders"],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Pedidos listados com sucesso"
          },
          "400": {
            "description": "Erro ao listar pedidos"
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customer_id": {
                    "type": "string"
                  },
                  "order_items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "item_id": {
                          "type": "number"
                        },
                        "order_item_qtd": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              },
              "examples": {
                "example1": {
                  "summary": "Valores possíveis para o campo <item_id>",
                  "value": {
                    "1": "Bebida",
                    "2": "Lanche",
                    "3": "Acompanhamento",
                    "4": "Sobremesa"
                  }
                },
                "example2": {
                  "summary": "Exemplo de corpo de requisição sem identificação do Cliente",
                  "value": {
                    "order_items": [
                      {
                        "order_item_qtd": 1,
                        "item_id": 1
                      }
                    ]
                  }
                },
                "example3": {
                  "summary": "Exemplo de corpo de requisição com Cliente identificado",
                  "value": {
                    "customer_id": 1,
                    "order_items": [
                      {
                        "order_item_qtd": 1,
                        "item_id": 1
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pedido criado com sucesso"
          },
          "400": {
            "description": "Erro ao criar pedido"
          }
        }
      }
    },
...
  }
}